Update test_cmd_up to include tests for "stack up -s"

+++ tests/test_cmd_up.py
@@ -72,11 +72,6 @@ class TestStackUp(unittest.TestCase):
         """
         Remove the temp directory and files
         """
-        my_vm_connection = vagrant_utils.Connect_to_vagrant(vm_name='rhel7-001',
-                                                            path=self.ctx.path)
-        ispoweron, isremote = vagrant_utils.vm_isrunning('rhel7-001', self.ctx.path)
-        if ispoweron == 0:
-            my_vm_connection.v.destroy(vm_name='rhel7-001')
         for f in os.listdir(self.hosts_path):
             file_name = os.path.join(self.hosts_path, f)
             os.remove(file_name)
@@ -87,15 +82,17 @@ class TestStackUp(unittest.TestCase):
         shutil.rmtree(self.backup_path)
         if hasattr(self, 'vagrant_bak'):
             os.rename(self.vagrant_bak, self.vagrant_yaml)
-        else:
+        elif os.path.isfile(self.vagrant_yaml):
             os.remove(self.vagrant_yaml)
         if hasattr(self, 'Vagrant_bak'):
             os.rename(self.Vagrant_bak, self.Vagrant_file)
-        else:
+        elif os.path.isfile(self.Vagrant_file):
             os.remove(self.Vagrant_file)
-        shutil.rmtree(self.dotvagrant_dir)
+        if os.path.isdir(self.dotvagrant_dir):
+            shutil.rmtree(self.dotvagrant_dir)
         if hasattr(self, 'dotvagrant_bak'):
             os.rename(self.dotvagrant_bak, self.dotvagrant_dir)
+        self.host = ''
         os.environ['OS_USERNAME'] = ''
         os.environ['OS_PASSWORD'] = ''
         os.environ['OS_REGION_NAME'] = ''
@@ -103,52 +100,125 @@ class TestStackUp(unittest.TestCase):
         os.environ['OS_TENANT_NAME'] = ''
         os.environ['OS_TENANT_ID'] = ''
 
-    def test_cmd_up_local_rhel7(self):
+    def cmd_up_runner(self, args, hostname, group, remote):
+        """
+        Run the 'stack up' command for each of the various tests
+
+        Args:
+            args {list}: CLI args
+            hostname {str}: Name of the vm
+            group {str}: Name of the group the vm belongs to
+            remote {bool}: True - OpenStack hypervisor
+                           False - VirtualBox hypervisor
+
+        Returns:
+            Nothing.  Runs tests based on the environment setup and updates logs directly
+
+        Example Usage:
+            self.cmd_up_runner(['--rhel7'], 'rhel7-001', False)
+        """
+        try:
+            runner = CliRunner()
+            host_yaml = hostname + '.yaml'
+            result = runner.invoke(cmd_up.cli, args)
+            vm_yaml_file = os.path.join(self.hosts_path, host_yaml)
+            self.assertTrue(os.path.isfile(vm_yaml_file))
+            self.ctx.logger.info('%s was found in dev-tenant-1/hosts.d' % host_yaml)
+            with open(vm_yaml_file, 'r') as yaml_file:
+                vm_yaml_data = yaml.load(yaml_file)
+            self.assertEqual(vm_yaml_data['groups'][1], group)
+            self.ctx.logger.info('%s VM group is %s as expected' % (hostname, group))
+            with open(self.vagrant_yaml, 'r') as vagrant_f:
+                vagrant_data = yaml.load(vagrant_f)
+            self.assertEqual(vm_yaml_data['interfaces']['eth0']['ip_address'],
+                             vagrant_data['hosts'][hostname]['ip'],
+                             '192.168.100.15')
+            self.ctx.logger.info('%s and vagrant.yaml have the expected IP' % host_yaml)
+            if remote:
+                hypervisor = "OpenStack"
+            else:
+                hypervisor = "VirtualBox"
+            ispoweron, isremote = vagrant_utils.vm_isrunning(hostname, self.ctx.path)
+            if ispoweron > 1:
+                self.ctx.logger.info('Unable to contact %s for VM status' % hypervisor)
+            elif ispoweron == 1:
+                self.ctx.logger.info('VM is offline in %s' % hypervisor)
+            else:
+                self.assertEqual(ispoweron, 0)
+                self.assertEqual(isremote, remote)
+                self.ctx.logger.info('VM is running on %s' % hypervisor)
+        except:
+            self.ctx.logger.debug('An unexpected error occured while processing command')
+
+    def destroy_vm(self, hostname):
+        """
+        Destroys vms based on the Vagrantfile data
+
+        Args:
+            hostname {str}: Name of the vm
+
+        Returns:
+            Nothing.  Destroys the VM within the hypervisor
+
+        Example Usage:
+            self.destroy_vm('rhel7-001')
+        """
+        my_vm_connection = vagrant_utils.Connect_to_vagrant(vm_name=hostname,
+                                                            path=self.ctx.path)
+        ispoweron, isremote = vagrant_utils.vm_isrunning(hostname, self.ctx.path)
+        # ispoweron states:
+        # 0 - VM is online
+        # 1 - VM is offline
+        # 2 - Vagrant command error, usually cannot find vm or contact hypervisor
+        # 3 - Other errors
+        if ispoweron <= 1:
+            my_vm_connection.v.destroy(vm_name=hostname)
+
+    def test_local_rhel7(self):
         """
         Tests the 'stack up --rhel7' command
         """
-        runner = CliRunner()
-        result = runner.invoke(cmd_up.cli, ['--rhel7'])
-        rhel_vm_file = os.path.join(self.hosts_path, 'rhel7-001.yaml')
-        self.assertTrue(os.path.isfile(rhel_vm_file))
-        self.ctx.logger.info('rhel7-001.yaml was found in dev-tenant-1/hosts.d')
-        with open(rhel_vm_file, 'r') as vm_file:
-            rhel_yaml = yaml.load(vm_file)
-        self.assertEqual(rhel_yaml['groups'][1], 'rhel7')
-        self.ctx.logger.info('Rhel7 VM group is rhel7 as expected')
-        with open(self.vagrant_yaml, 'r') as vagrant_f:
-            vagrant_yaml = yaml.load(vagrant_f)
-        self.assertEqual(rhel_yaml['interfaces']['eth0']['ip_address'],
-                         vagrant_yaml['hosts']['rhel7-001']['ip'])
-        self.ctx.logger.info('rhel7-001.yaml and vagrant.yaml have the same IP')
+        args = ['--rhel7']
+        hostname = 'rhel7-001'
+        group = 'rhel7'
+        remote = False
+        self.cmd_up_runner(args, hostname, group, remote)
+        self.destroy_vm(hostname)
 
-    def test_cmd_up_remote_rhel7(self):
+    def test_remote_rhel7(self):
         """
         Tests the 'stack up --rhel7 -r' command
         """
-        runner = CliRunner()
-        result = runner.invoke(cmd_up.cli, ['--rhel7', '-r'])
-        rhel_vm_file = os.path.join(self.hosts_path, 'rhel7-001.yaml')
-        self.assertTrue(os.path.isfile(rhel_vm_file))
-        self.ctx.logger.info('rhel7-001.yaml was found in dev-tenant-1/hosts.d')
-        with open(rhel_vm_file, 'r') as vm_file:
-            rhel_yaml = yaml.load(vm_file)
-        self.assertEqual(rhel_yaml['groups'][1], 'rhel7')
-        self.ctx.logger.info('Rhel7 VM group is %s as expected' % rhel_yaml['groups'][1])
-        with open(self.vagrant_yaml, 'r') as vagrant_f:
-            vagrant_yaml = yaml.load(vagrant_f)
-        self.assertEqual(rhel_yaml['interfaces']['eth0']['ip_address'],
-                         vagrant_yaml['hosts']['rhel7-001']['ip'],
-                         '192.168.100.15')
-        self.ctx.logger.info('rhel7-001.yaml and vagrant.yaml have the same IP')
-        ispoweron, isremote = vagrant_utils.vm_isrunning('rhel7-001', self.ctx.path)
-        if ispoweron > 1:
-            self.ctx.logger.info('Unable to contact Openstack provider for VM status')
-        else:
-            self.assertEqual(ispoweron, 0)
-            self.ctx.logger.info('VM is running')
-            self.assertEqual(isremote, True)
-            self.ctx.logger.info('VM is remotely installed')
+        args = ['--rhel7', '-r']
+        hostname = 'rhel7-001'
+        group = 'rhel7'
+        remote = True
+        self.cmd_up_runner(args, hostname, group, remote)
+        self.destroy_vm(hostname)
+
+    def test_local_service(self):
+        """
+        Tests the 'stack up -s <servicename>' command
+        """
+        args = ['-s', 'service-sonarqube']
+        hostname = 'service-sonarqube-001'
+        group = 'sonarqube'
+        remote = False
+        self.cmd_up_runner(args, hostname, group, remote)
+        self.destroy_vm(hostname)
+        self.destroy_vm('infra-001')
+
+    def test_remote_service(self):
+        """
+        Tests the 'stack up -s <servicename> -r' comman
+        """
+        args = ['-s', 'service-sonarqube', '-r']
+        hostname = 'service-sonarqube-001'
+        group = 'sonarqube'
+        remote = True
+        self.cmd_up_runner(args, hostname, group, remote)
+        self.destroy_vm(hostname)
+        self.destroy_vm('infra-001')
 
 
 if __name__ == '__main__':
